+++ app/routes/admin.drinks.new.tsx
@@ -1,5 +1,5 @@
 import { redirect, href, data } from "react-router";
-import { getSession, commitSession } from "#/app/auth/session.server";
+import { getSession, commitSession } from "#/app/modules/auth";
 import { createDrink } from "#/app/models/drink.server";
 import { uploadImageOrPlaceholder } from "#/app/utils/imagekit.server";
 import { DrinkForm } from "#/app/admin/drink-form";
